Proxy changes due to fraudulents domains

+++ src/core/controllers/CampaignsController.php
@@ -191,7 +191,7 @@
       $this->setViewData('campaign', $campaign);
 
       // Rozmazání identit uživatelů dle konfigurace.
-      $this->setViewData('blurIdentities', (((int) CAMPAIGN_STATS_BLUR_IDENTITIES) ? 'blur-text' : ''));
+      $this->setViewData('blurIdentities', (CAMPAIGN_STATS_BLUR_IDENTITIES ? 'blur-text' : ''));
 
       // Získání nasbíraných dat.
       if (isset($_GET[ACT_STATS_ALL_ACTIONS])) {
